requirements.txt
Flask
PyGithub